pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shopee-affiliate-agent"
version = "0.1.0"
description = "Shopee affiliate to Facebook content operations agent"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Codex" }
]
dependencies = [
  "openai>=1.12.0",
  "apscheduler>=3.10.4",
  "chromadb>=1.0.8",
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "langgraph>=0.2.74",
  "pillow>=11.0.0",
  "playwright>=1.50.0",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "redis>=5.2.1",
  "rq>=2.1.0",
  "sse-starlette>=2.1.3",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3"
]

[tool.setuptools]
packages = ["common", "core", "modules", "modules.shopee", "modules.ai", "modules.meta", "modules.memory", "web"]
include-package-data = true

[tool.setuptools.package-data]
web = ["templates/*.html", "static/*.css"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
